fix codegen

+++ frontend/src/graphql/types.ts
@@ -2683,9 +2683,9 @@ export type LogoutResult = {
 /** The Me type represents the currently authenticated user. */
 export type Me = {
   __typename?: 'Me';
-  assistantMonthlyCost: Scalars['Float']['output'];
+  assistantMonthlyCost: Scalars['Int']['output'];
   assistantMonthlyLimitExceeded: Scalars['Boolean']['output'];
-  assistantTotalCost: Scalars['Float']['output'];
+  assistantTotalCost: Scalars['Int']['output'];
   /** The feature flags assigned to the currently authenticated user. */
   features: Array<FeatureFlag>;
   hasTwoFactorEnabled: Scalars['Boolean']['output'];
